refactor: break monolithic updater apart and move out library parts into separate package
chore: explicit ignore of return vals
